Updated summing of two list
You have two numbers represented by a linked list, where each node contains a single digit. The digits are stored in reverse order, such that the 1's digit is at the head of the list. Write a function that adds the two numbers and returns the sum as a linked list.

list1 = 7 -> 1 -> 6
list2 =  5 -> 9 -> 2
result = 2 -> 1 -> 9